+++ mqttwarn/__init__.py
@@ -7,7 +7,7 @@ __license__ = "Eclipse Public License - v 2.0 (http://www.eclipse.org/legal/epl-
 
 try:
     from importlib.metadata import version
-except ImportError:
+except ImportError:  # pragma: nocover
     from importlib_metadata import version  # type: ignore[no-redef]
 
 __version__ = version("mqttwarn")